+++ source/source_pw/module_pwdft/dftu_base_io.h
@@ -2,7 +2,7 @@
 #define DFTU_BASE_IO_H
 
 #include "source_base/matrix.h"
-#include "source_pw/module_pwdft/occ_matrix.h"
+#include "source_estate/occ_matrix.h"
 
 #include <iosfwd>
 #include <string>
